yesdk-804 get assertion command (#509)
* start MakeCredentialParameters class

* add more Make Credential parameters

* fix merge issues

* fixed error in test

* update based on review

* fixed some issues

* add first draft of Make Credential Command

* update based on review

* more make credential response decoding

* add code to parse and process the MakeCredential response

* add GetAssertionParameters and update Cbor code

* update based on review

* fix merge issues and initial code for GetAssertionCommand

* one more merge fix

* finish GetAssertionCOmmand and its response

* Fix error in class name

+++ Yubico.YubiKey/src/Yubico/YubiKey/Fido2/Commands/ClientPinCommand.cs
@@ -166,14 +166,14 @@ namespace Yubico.YubiKey.Fido2.Commands
         {
             var cbor = new CborWriter(CborConformanceMode.Ctap2Canonical, convertIndefiniteLengthEncodings: true);
 
-            CborHelpers.BeginMap<long>(cbor)
-                .OptionalEntry(TagPinUvAuthProtocol, (uint?)PinUvAuthProtocol)
-                .Entry(TagSubCommand, (uint)SubCommand)
+            CborHelpers.BeginMap<int>(cbor)
+                .OptionalEntry(TagPinUvAuthProtocol, (int?)PinUvAuthProtocol)
+                .Entry(TagSubCommand, (int)SubCommand)
                 .OptionalEntry(TagKeyAgreement, KeyAgreement)
                 .OptionalEntry(TagPinUvAuthParam, PinUvAuthParam)
                 .OptionalEntry(TagNewPinEnc, NewPinEnc)
                 .OptionalEntry(TagPinHashEnc, PinHashEnc)
-                .OptionalEntry(TagPermissions, (uint?)Permissions)
+                .OptionalEntry(TagPermissions, (int?)Permissions)
                 .OptionalEntry(TagRpId, RpId)
                 .EndMap();
